Implement Warning the user after successful added if the series was 100 before to reach the maximum 10 digits SeriesNumber in SI

+++ Repository/SalesInvoiceRepo.cs
@@ -21,7 +21,7 @@ namespace Accounting_System.Repository
                 .ToListAsync();
         }
 
-        public async Task<int> GetLastSeriesNumber()
+        public async Task<long> GetLastSeriesNumber()
         {
             var lastInvoice = await _dbContext
                 .SalesInvoices
